requirements/dev.txt
-r common.txt

django-debug-toolbar==1.6
watchdog==0.8.3
